requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
allure-pytest==2.12.0
allure-python-commons==2.12.0
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.1.0
black==23.12.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
devtools==0.12.2
executing==2.0.1
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
isort==5.13.2
kaitaistruct==0.10
mimesis==11.1.0
multidict==6.0.4
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic-settings==2.1.0
pydantic_core==2.14.3
pyflakes==3.1.0
Pygments==2.17.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.2.1
pytest-rerunfailures==13.0
python-dotenv==0.21.1
requests==2.31.0
selenium==4.8.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==1.26.18
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
zstandard==0.22.0
